fix(ETL): use 0.001 default values to ensure QuantumLeap creates DOUBLE PRECISION columns
- Replace all 0.0 defaults with 0.001 for pollutants (CO, NO, NO2, O3, SO2, PM2.5, PM10, NH3)
- Update precipitation and pressureTendency defaults to 0.001
- Remove int*10 conversion, use direct float values with round(x, 2)
- Prevent TimescaleDB type inference from creating BIGINT instead of DOUBLE PRECISION

+++ backend/src/main/java/org/opensource/smartair/dtos/WeatherDataDTO.java
@@ -69,8 +69,8 @@ public class WeatherDataDTO {
     private Integer cloudiness; // 0-100 %
     private Integer illuminance; // lux
 
-    // Optional
-    private Integer pressureTendency;
+    // Optional (converted from double, 0.001 → 0)
+    private Double pressureTendency; // hPa/h
 
     @JsonProperty("refDevice")
     private String refDevice; // Device URN
